now finish

+++ server/controllers/posts.controller.js
@@ -32,20 +32,29 @@ function getById(req, res) {
 }
 
 function update(req, res) {
-    postService.update(req.params._id, req.body)
-        .then(() => {
-            userService.updateTaggedUsers(req.params._id, req.body.toAdd, req.body.toRemove)
+    postService.getById(req.params._id).then(post => {
+        if (req.user.sub == post.userId) {
+            postService.update(req.params._id, req.body)
                 .then(() => {
-                    res.sendStatus(200);
+                    userService.updateTaggedUsers(req.params._id, req.body.toAdd, req.body.toRemove)
+                        .then(() => {
+                            res.sendStatus(200);
+                        }).catch(err => {
+                            res.status(400).send(err);
+                        });
                 }).catch(err => {
                     res.status(400).send(err);
                 });
-        }).catch(err => {
-            res.status(400).send(err);
-        });
+        } else {
+            res.status(401).send("Unauthorized");
+        }
+    }).catch(err => {
+        res.status(400).send(err);
+    });
+
 }
 
-function search(req, res){
+function search(req, res) {
     let query = req.body.query;
     postService.search(query)
         .then((posts) => {
@@ -71,46 +80,58 @@ function getPopular(req, res) {
 }
 
 function comment(req, res) {
-    postId = req.params._id;
-    comment = req.body;
-    postService.addComment(postId, comment)
-        .then(() => {
-            res.sendStatus(200);
-        }).catch(err => {
-            res.status(400).send(err);
-        });
+    if (req.user.sub == req.body.userId) {
+        postId = req.params._id;
+        comment = req.body;
+        postService.addComment(postId, comment)
+            .then(() => {
+                res.sendStatus(200);
+            }).catch(err => {
+                res.status(400).send(err);
+            });
+    } else {
+        res.status(401).send("Unauthorized");
+    }
 }
 
 function like(req, res) {
     user = req.body;
     postId = req.params._id;
-    postService.addLike(postId, user._id)
-        .then(() => {
-            userService.addToLikes(user._id)
-                .then(() => {
-                    res.sendStatus(200);
-                }).catch(err => {
-                    res.status(400).send(err);
-                });
-        }).catch(err => {
-            res.status(400).send(err);
-        });
+    if (req.user.sub == user._id) {
+        postService.addLike(postId, user._id)
+            .then(() => {
+                userService.addToLikes(user._id)
+                    .then(() => {
+                        res.sendStatus(200);
+                    }).catch(err => {
+                        res.status(400).send(err);
+                    });
+            }).catch(err => {
+                res.status(400).send(err);
+            });
+    } else {
+        res.status(401).send("Unauthorized");
+    }
 }
 
 function unlike(req, res) {
     user = req.body;
     postId = req.params._id;
-    postService.removeLike(postId, user._id)
-        .then(() => {
-            userService.removeFromLikes(user._id)
-                .then(() => {
-                    res.sendStatus(200);
-                }).catch(err => {
-                    res.status(400).send(err);
-                });
-        }).catch(err => {
-            res.status(400).send(err);
-        });
+    if (req.user.sub == user._id) {
+        postService.removeLike(postId, user._id)
+            .then(() => {
+                userService.removeFromLikes(user._id)
+                    .then(() => {
+                        res.sendStatus(200);
+                    }).catch(err => {
+                        res.status(400).send(err);
+                    });
+            }).catch(err => {
+                res.status(400).send(err);
+            });
+    } else {
+        res.status(401).send("Unauthorized");
+    }
 }
 
 function getMembers(req, res) {
@@ -144,7 +165,6 @@ function feedPosts(req, res) {
     userService.getById(params._id)
         .then((user) => {
             var userList = user.following;
-
             userList.push(user._id);
             postService.getPosts(parseInt(params.startIndex), parseInt(params.limit), userList, user.taggedPosts)
                 .then((posts) => {
@@ -159,21 +179,27 @@ function feedPosts(req, res) {
 }
 
 function newPost(req, res) {
-    postService.publish(req.body)
-        .then(() => {
-            post = req.body;
-            userService.addPostToUser(post.userId, post.taggedUsers, post._id)
-                .then(() => {
-                    res.sendStatus(200);
-                }).catch(err => {
-                    res.status(400).send(err);
-                });
-        }).catch((err) => {
-            res.status(400).send(err);
-        });
+    if (req.user.sub == req.body.userId) {
+        postService.publish(req.body)
+            .then(() => {
+                post = req.body;
+                userService.addPostToUser(post.userId, post.taggedUsers, post._id)
+                    .then(() => {
+                        res.sendStatus(200);
+                    }).catch(err => {
+                        res.status(400).send(err);
+                    });
+            }).catch((err) => {
+                res.status(400).send(err);
+            });
+    } else {
+        res.status(401).send("Unauthorized");
+    }
 
 }
 
+
+//for debuging
 function getAll(req, res) {
     postService.getAll()
         .then(function (users) {
@@ -187,17 +213,21 @@ function getAll(req, res) {
 function _delete(req, res) {
     postService.getById(req.params._id)
         .then((post) => {
-            userService.removePostFromUser(post.userId, post.taggedUsers, post._id)
-                .then(() => {
-                    postService.delete(req.params._id)
-                        .then(() => {
-                            res.sendStatus(200);
-                        }).catch(err => {
-                            res.status(400).send(err);
-                        });
-                }).catch((err) => {
-                    res.status(400).send(err);
-                });
+            if (post.userId == req.user.sub) {
+                userService.removePostFromUser(post.userId, post.taggedUsers, post._id)
+                    .then(() => {
+                        postService.delete(req.params._id)
+                            .then(() => {
+                                res.sendStatus(200);
+                            }).catch(err => {
+                                res.status(400).send(err);
+                            });
+                    }).catch((err) => {
+                        res.status(400).send(err);
+                    });
+            } else {
+                res.status(401).send("Unauthorized");
+            }
         }).catch((err) => {
             res.status(400).send(err);
         });
